refactored redux student package
