+++ web_front_end/src/containers/Home/TableDataGrid/TableDataGrid.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render, screen } from "@testing-library/react";
 import { Provider } from "react-redux";
 import store from "../../../redux/store";
 import TableDataGrid from "./TableDataGrid";
-import { tableDataActions } from "../../../redux/tableSlice/tableSlice";
+import { tableDataActions } from "../../../redux/student/studentSlice";
 import { maxDate } from "../../../util/validateTable";
 
 interface ITableData {
